Add files via upload
This is the database file for the project.  Copy the file to a folder named 'db' in 'C drive' and use the url 'jdbc:h2:tcp://localhost//C:/db/h2DS' for connecting to the H2 database.  The database should be connect during compiling the project and executing the JAR file.